requirements.txt
astroid==2.5.6
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
decorator==5.0.9
Flask==2.0.1
greenlet==1.1.0
idna==2.10
inflect==5.3.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
-e git+https://github.com/biddy1618/newsProject.git@902dbcc75c31b34cb3f9bb955f0126138f77ad9f#egg=projectNews
prompt-toolkit==3.0.18
psycopg2==2.8.6
ptyprocess==0.7.0
Pygments==2.9.0
pylint==2.8.3
python-dateutil==2.8.1
python-dotenv==0.17.1
pyzmq==22.1.0
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlacodegen==2.3.0
SQLAlchemy==1.4.18
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
